iam/requirements.txt
grpcio==1.80.0
grpcio-tools==1.80.0
asyncpg==0.31.0
alembic==1.18.4
pyjwt==2.12.1
structlog==25.5.0
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0
redis==7.4.0

passlib[bcrypt]
sqlalchemy[asyncio]
